Исправляет ошибки линтера
